+++ src/scripts/objects/debug/fps-text/fpsText.ts
