+++ src/scripts/objects/debug/index.ts
@@ -1,1 +1,1 @@
-export { default as FpsText } from './fpsText'
+export { default as FpsText } from './fps-text/fpsText'
