+++ src/scripts/objects/gui/game-ui/index.ts
@@ -1,0 +1,1 @@
+export { default as PhaserLogo } from './phaser-logo/phaserLogo'
