+++ src/scripts/objects/gui/game-ui/phaser-logo/phaserLogo.ts
