--- src/scripts/objects/gui/index.ts
@@ -1,1 +1,0 @@
-export { default as PhaserLogo } from './phaserLogo'
